requirements.txt
python-owasp-zap-v2.4==0.0.20
requests==2.31.0
schedule==1.2.0
jinja2==3.1.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
pyyaml==6.0.1
docker==7.0.0
matplotlib>=3.5.0
pandas>=1.3.0